requirements.txt
streamlit
google-generativeai
docx
PyPDF2
